# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: CS4783Assignment
  description: Jason Kha, and Jennifer Nguyen
host: localhost:3000
basePath: /
schemes:
  - https
tags:
  - name: Hello
    description: Hello
  - name: Property
    description: Property

paths:
  /hello:
    get:
      description: OK
      tags:
        - Hello
      responses:
        "200":
          description: OK
          schema:
            $ref: "../swagger/definitions.yaml#/definitions/msg"

  /properties:
    post:
      summary: Add a Property
      tags:
        - Property
      operationId: postProperty
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: property
          description: The property to create
          schema:
            $ref: "../swagger/definitions.yaml#/definitions/POSTProperty"
      responses:
        "200":
          description: Successfully Created a Property
          schema:
            type: object
            items:
              $ref: "../swagger/definitions.yaml#/definitions/msg"
            example:
              message: added
        "400":
          description: Error
          schema:
            type: array
            items:
              $ref: "../swagger/definitions.yaml#/definitions/msg"
            example:
              - message: address is not between 1 and 200 characters
              - message: state is not 2 characters
              - message: city is not between 1 and 50 characters
              - message: zip is not between 5 and 10 characters

    get:
      summary: Return a list properties
      description: Returns a list of Stored US properties (id,address, and zip) order by id
      tags:
        - Property
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "../swagger/definitions.yaml#/definitions/GETProperty"
            example:
              - id: 1
                address: 123 Test Ave
                zip: 78222
              - id: 2
                address: 123 Main Street
                zip: 78222
  /properties/{propertyId}:
    get:
      summary: Find a Property
      description: return a JSON
      tags:
        - Property
      parameters:
        - in: path
          name: propertyId
          type: integer
          required: true
          description: Numberic ID of the property to get
      responses:
        "200":
          description: Successfully Retrived a Property
          schema:
            $ref: "../swagger/definitions.yaml#/definitions/property"
            example:
              id: 1
              address: 123 Test Ave
              city: San Antonio
              state: TX
              zip: 78222
        "404":
          description: Property not found

    delete:
      description: Delete the property with an id value {id}
      tags:
        - Property
      responses:
        "200":
          description: OK